Revert as fix for alpha pdf backgrounds didn't work

+++ app/models/concerns/file_behavior.rb
@@ -48,12 +48,14 @@ module FileBehavior
     self.set_parent_pid
     self.attempt_initialize!
 
+    unless File.exists?(full_path)
     #unless File.exists?(full_path) and !self.class.name.include?("File")
       FileUtils.mkpath self.full_directory
       File.open(self.full_path, 'wb' ) do |output|
         output.write value
       end
     #end
+    end
   end
 
   def content
